feat(usecase): moved Logo to admin components + updated yarn lock
