+++ packages/core/admin/admin/src/pages/AuthPage/components/Oops/index.js
@@ -10,7 +10,7 @@ import UnauthenticatedLayout, {
   Column,
   LayoutContent,
 } from '../../../../layouts/UnauthenticatedLayout';
-import Logo from '../Logo';
+import Logo from '../../../../components/UnauthenticatedLogo';
 
 const Oops = () => {
   const { formatMessage } = useIntl();
