+++ packages/core/admin/admin/src/pages/AuthPage/components/Register/index.js
@@ -24,7 +24,7 @@ import UnauthenticatedLayout, {
   Column,
   LayoutContent,
 } from '../../../../layouts/UnauthenticatedLayout';
-import Logo from '../Logo';
+import Logo from '../../../../components/UnauthenticatedLogo';
 import FieldActionWrapper from '../FieldActionWrapper';
 
 const CenteredBox = styled(Box)`
